+++ packages/core/src/enums/VideoViewport.ts
@@ -1,0 +1,9 @@
+/**
+ * video viewport speed units
+ */
+enum SpeedUnit {
+  FRAME = 'f',
+  SECOND = 's',
+}
+
+export { SpeedUnit };