+++ packages/tools/examples/volumeViewportOrientation/index.ts
@@ -9,6 +9,8 @@ import {
   StackScrollMouseWheelTool,
   ToolGroupManager,
   addTool,
+  Enums as csToolsEnums,
+  ZoomTool,
 } from '@cornerstonejs/tools';
 import {
   initDemo,
@@ -16,6 +18,7 @@ import {
   setTitleAndDescription,
   addDropdownToToolbar,
 } from '../../../../utils/demo/helpers';
+const { MouseBindings } = csToolsEnums;
 
 // This is for debugging purposes
 console.warn(
@@ -43,14 +46,14 @@ const element = document.createElement('div');
 element.id = 'cornerstone-element';
 element.style.width = '500px';
 element.style.height = '500px';
-
+element.oncontextmenu = () => false;
 content.appendChild(element);
 // ============================= //
 
 addDropdownToToolbar({
   options: {
     values: ['axial', 'sagittal', 'coronal', 'acquisition'],
-    defaultValue: 'axial',
+    defaultValue: 'sagittal',
   },
   onSelectedValueChange: (selectedValue) => {
     // Get the rendering engine
@@ -72,22 +75,31 @@ async function run() {
   // Init Cornerstone and related libraries
   await initDemo();
   addTool(StackScrollMouseWheelTool);
+  addTool(ZoomTool);
 
   // Using a oblique acquired image to demonstrate the orientation of the volume
   // in default (acquisition plane mode)
   const imageIds = await createImageIdsAndCacheMetaData({
     StudyInstanceUID:
-      '1.3.6.1.4.1.14519.5.2.1.7311.5101.247316591887822227457894627822',
+      '1.3.6.1.4.1.14519.5.2.1.7009.2403.871108593056125491804754960339',
     SeriesInstanceUID:
-      '1.3.6.1.4.1.14519.5.2.1.7311.5101.119114186762760923175160291330',
-    wadoRsRoot: 'https://server.dcmjs.org/dcm4chee-arc/aets/DCM4CHEE/rs',
+      '1.3.6.1.4.1.14519.5.2.1.7009.2403.367700692008930469189923116409',
+    wadoRsRoot: 'https://d33do7qe4w26qo.cloudfront.net/dicomweb',
   });
 
   // create toolGroup
   const toolGroup = ToolGroupManager.createToolGroup('myToolGroup');
   toolGroup.addTool(StackScrollMouseWheelTool.toolName);
+  toolGroup.addTool(ZoomTool.toolName);
 
   toolGroup.setToolActive(StackScrollMouseWheelTool.toolName);
+  toolGroup.setToolActive(ZoomTool.toolName, {
+    bindings: [
+      {
+        mouseButton: MouseBindings.Secondary,
+      },
+    ],
+  });
   toolGroup.addViewport(viewportId, renderingEngineId);
 
   // Instantiate a rendering engine
@@ -99,7 +111,7 @@ async function run() {
     type: ViewportType.ORTHOGRAPHIC,
     element,
     defaultOptions: {
-      orientation: Enums.OrientationAxis.AXIAL,
+      orientation: Enums.OrientationAxis.SAGITTAL,
       background: <Types.Point3>[0.2, 0, 0.2],
     },
   };