+++ packages/tools/src/tools/OrientationMarkerTool.ts
@@ -94,6 +94,36 @@ class OrientationMarkerTool extends BaseTool {
     this.initViewports();
   };
 
+  onSetToolDisabled = (): void => {
+    this.cleanUpData();
+  };
+
+  private cleanUpData() {
+    const renderingEngines = getRenderingEngines();
+    const renderingEngine = renderingEngines[0];
+    const viewports = renderingEngine.getViewports();
+
+    viewports.forEach((viewport) => {
+      const orientationMarker = this.orientationMarkers[viewport.id];
+      if (!orientationMarker) {
+        return;
+      }
+
+      const { actor, orientationWidget } = orientationMarker;
+      orientationWidget?.setEnabled(false);
+      orientationWidget?.delete();
+      actor?.delete();
+
+      const renderWindow = viewport
+        .getRenderingEngine()
+        .offscreenMultiRenderWindow.getRenderWindow();
+      renderWindow.render();
+      viewport.getRenderingEngine().render();
+
+      delete this.orientationMarkers[viewport.id];
+    });
+  }
+
   private initViewports() {
     const renderingEngines = getRenderingEngines();
     const renderingEngine = renderingEngines[0];
@@ -159,7 +189,6 @@ class OrientationMarkerTool extends BaseTool {
       orientationWidget,
       actor,
     };
-    renderer.resetCamera();
     renderWindow.render();
     viewport.getRenderingEngine().render();