+++ packages/tools/src/utilities/index.ts
@@ -9,6 +9,8 @@ import throttle from './throttle';
 import isObject from './isObject';
 import clip from './clip';
 import calibrateImageSpacing from './calibrateImageSpacing';
+import getCalibratedLengthUnits from './getCalibratedUnits';
+import { getCalibratedScale } from './getCalibratedUnits';
 import triggerAnnotationRenderForViewportIds from './triggerAnnotationRenderForViewportIds';
 import triggerAnnotationRender from './triggerAnnotationRender';
 import jumpToSlice from './viewport/jumpToSlice';
@@ -35,6 +37,7 @@ import * as viewport from './viewport';
 import * as touch from './touch';
 import * as dynamicVolume from './dynamicVolume';
 import * as polyDataUtils from './polyData/utils';
+import * as voi from './voi';
 
 // Events
 import { triggerEvent } from '@cornerstonejs/core';
@@ -52,6 +55,8 @@ export {
   touch,
   triggerEvent,
   calibrateImageSpacing,
+  getCalibratedLengthUnits,
+  getCalibratedScale,
   segmentation,
   triggerAnnotationRenderForViewportIds,
   triggerAnnotationRender,
@@ -72,4 +77,5 @@ export {
   roundNumber,
   pointToString,
   polyDataUtils,
+  voi,
 };