+++ packages/tools/src/utilities/voi/colorbar/types/ColorbarSize.ts
@@ -1,0 +1,4 @@
+export type ColorbarSize = {
+  width: number;
+  height: number;
+};